(WIP) Adding fish support
This PR adds fish[1] support to shell-hist. Unfortunately it seems to
still be slightly buggy, as the generated statics don't really make
much sense - maybe you have an idea of what's going on here.

But generally: fish's `fish_history` is kinda annoying to parse since
they use an almost(tm)-yaml syntax. So instead I restructured the code
to act on a large string instead, then providing a function that
returns said string. In the case of `fish` we just shell out to
`history(1)` to read the `fish_history` for us :)